+++ ax3l/constants/DZMQ.py
@@ -4,3 +4,4 @@ from typing import Final
 class DZMQ:
     PROTOCOL_VERSION: Final[int] = 1
     TIMEOUT_SECONDS: Final[float] = 15.0
+    POLL_MS: Final[int] = 100